--- src/models/devicetype.go
@@ -1,6 +1,0 @@
-package models
-
-type DeviceType struct {
-	Id   uint64 `json:"id"`
-	Name string `json:"name"`
-}
