--- src/models/location.go
@@ -1,6 +1,0 @@
-package models
-
-type Location struct {
-	Id   uint64 `json:"id"`
-	Name string `json:"name"`
-}
