+++ src/restService/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_binary(
+    name = "restService",
+    embed = [":restService_lib"],
+    visibility = ["//visibility:public"],
+)
+
+go_library(
+    name = "restService_lib",
+    srcs = ["main.go"],
+    importpath = "mstraubAC/smarthome-restService/restService",
+    visibility = ["//visibility:private"],
+    deps = [
+        "@com_github_gin_gonic_gin//:go_default_library",
+        "@org_uber_go_zap//:go_default_library",
+    ],
+)
