+++ src/restService/accessors/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "accessors",
+    srcs = ["database.go"],
+    importpath = "mstraubAC/smarthome-restService/restService/accessors",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_jackc_pgx_v5//pgxpool:go_default_library",
+        "@org_uber_go_zap//:go_default_library",
+    ],
+)
