+++ src/restService/controllers/locations/BUILD.bazel
@@ -1,0 +1,17 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "locations",
+    srcs = [
+        "controllers.go",
+        "get_location.go",
+        "get_locations.go",
+    ],
+    importpath = "mstraubAC/smarthome-restService/restService/controllers/locations",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_georgysavva_scany_v2//pgxscan:go_default_library",
+        "@com_github_gin_gonic_gin//:go_default_library",
+        "@org_uber_go_zap//:go_default_library",
+    ],
+)
