+++ src/restService/middleware/ginlogging.go
@@ -1,0 +1,49 @@
+package middleware
+
+import (
+	"bytes"
+	"io"
+	"time"
+
+	ginzap "github.com/gin-contrib/zap"
+	"github.com/gin-gonic/gin"
+	"go.opentelemetry.io/otel/trace"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+)
+
+func ZapLoggingHandler(logger *zap.Logger) gin.HandlerFunc {
+	return ginzap.GinzapWithConfig(
+		logger,
+		&ginzap.Config{
+			UTC:        true,
+			TimeFormat: time.RFC3339,
+			Context:    ginzap.Fn(ZapContextLoggingSupporter),
+		},
+	)
+}
+
+func ZapContextLoggingSupporter(c *gin.Context) []zapcore.Field {
+	fields := []zapcore.Field{}
+
+	// log request ID
+	if requestID := c.Writer.Header().Get("X-Request-Id"); requestID != "" {
+		fields = append(fields, zap.String("request_id", requestID))
+	}
+
+	// log trace and span ID
+	if trace.SpanFromContext(c.Request.Context()).SpanContext().IsValid() {
+		fields = append(fields, zap.String("trace_id", trace.SpanFromContext(c.Request.Context()).SpanContext().TraceID().String()))
+		fields = append(fields, zap.String("span_id", trace.SpanFromContext(c.Request.Context()).SpanContext().SpanID().String()))
+	}
+
+	// log request body
+	var body []byte
+	var buf bytes.Buffer
+	tee := io.TeeReader(c.Request.Body, &buf)
+	body, _ = io.ReadAll(tee)
+	c.Request.Body = io.NopCloser(&buf)
+	fields = append(fields, zap.String("body", string(body)))
+
+	return fields
+}
