+++ src/restService/models/devicetype.go
@@ -1,0 +1,6 @@
+package models
+
+type DeviceType struct {
+	Id   uint64 `json:"id"`
+	Name string `json:"name"`
+}
