+++ src/restService/models/location.go
@@ -1,0 +1,6 @@
+package models
+
+type Location struct {
+	Id   uint64 `json:"id"`
+	Name string `json:"name"`
+}
